Update tests to use DMS proxies
Instead of relying on the Truffle migrations to initialize test context
implicitly, use the deployed ProxyFactory to create new DMS proxies
as-appropriate in the tests. This is how most users will get a DMS
contract anyway, so it is the path that we should test. We should also
reduce cross-test statefullness, as it makes individual tests harder to
reason about, so this removes the "if test" token deployment from the
migrations and, where appropriate, deploys a fresh token (and a fresh
DMS proxy) for each test.

* Use makeDMSProxy instead of Dead.deployed()
* Give each test fresh contracts (DMS + token) to reduce statefulness
* Update depositTokens utility to accept a token addr and DMS addr